Fix gauntlet shop transition and hide reserve text
